+++ PowerShell/Get-AzureStorageFileContent/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 1.0.1 - 2021-02-26
 
+    * Fixed error when folder already exists
+    * Added `Force` parameter to overwrite the existing file even if it has the same name and size. I can't think why this would be needed but I added it anyway.
+    * Added destination path to output
+
+## 1.0.1 - 2021-02-26
+
     * Fix files being always skipped.
 
 ## 1.0.0 - 2021-02-23